refactored styles, components, containers into their own folders
